PP Attempt, bullet FX, Masking, Damage taken FX, camera move FX
post processing is a bitch
bullet FX fixed
Mask on scene enter, and death
edge of screen flashes red when taking damage
camera zooms in on the players dead body because cinematic. (this is also kind of weird, and hopefully the camera is resetting it's position properly. But atm its weird so I've changed the BG to black to match the blindonator3001)